Consistently use db_prepare_bool() for boolean columns
In Version API, the released column was treated as an integer, despite
the fact that it's defined as 'L' in the schema. This can cause problems
with PostgreSQL.

Also fixes #17333, Oracle throwed an ORA-01400 error when adding a new
version (regression caused by a80f447687722b52160d00c96c21a5be5f4329e1,
which removed the call to db_prepare_bool()).