Update constants to be consistent with underlying db column type
We define constants VERSION_FUTURE = 0 and VERSION_RELEASED = 1
for use with project_version.released column. However, it is defined as
type 'L' which maps to boolean in some DB types.

To prevent errors caused by mismatched data types, the constants are now
using true/false. If we ever need to store more values, the column's
type should be changed to integer in the schema.