Push current plugin in plugin_needs_upgrade()
Without this, the plugins' schema() methods will return incorrect table
names (i.e. without the plugin's name as a prefix) if they rely on
plugin_table() to build them.

This is really picking nits, because an incorrect table name is very
unlikely to cause issues since the function only checks the number of
entries in the plugin's schema array.