Fix 'project not found' message from roadmap/changelog
The Roadmap and Change Log can receive indication of a project either as
an id or a name. In the latter case, if the project is not found the
error message printed "0" instead of the actual value of the project
parameter.

This commit fixes the error message to allow strings instead of decimal,
and adds error_parameters() calls to send the project name to the error
handler.

Fixes #17370