Do not register previously registered plugins
It may happen that a plugin listed as enabled in the plugin table is in
fact already registered with MantisBT if that plugin is also listed as
force_installed in the configuration.

In that case, re-registering it basically 'unprotects' it, and
manage_plugin_page.php displays an Uninstall action even though the
plugin can't actually be uninstalled.