Fix #0016570: Page content is forgotten when user clicks [Back] button
By default we disable $g_allow_browser_cache in all circumstances.

The config value is commented out in config_defaults.inc.php. In fact, thraxisp
added the config option in 2005 in the commented out state.

It would appear the DHX re-ordered this call when implemented require_api
- see:

 https://github.com/mantisbt/mantisbt/commit/2d5455cce290531359a55d353010258c6f3f6f34

The above code is only in 1.3 and not in the 1.2 series, so the above fix should
be safe to apply.

We probably need to consider whether we actually want to remove this configuration option

After all, it's a configuration option we've probably never had anyone use in 5 years
so I expect that the value of it is rather....