Fix absolute URLs issue with 'User Documentation' link
This resolves the issue with absolute URLs specified in $g_manual_url
becoming relative on project documentation pages.

Following discussions in the pull request [1], the solution implements
additional logic in the print_doc_menu() function that checks whether
the provided URL is absolute or relative (based on presence of an URL
scheme):
- if relative, makes sure the local path exists
  - if yes, build absolute URL using helper_mantis_url()
  - if not, fall back to the online manual on
    http://www.mantisbt.org/documentation.php
- if absolute, uses the URL as-is without check

Since this is currently the only use case for absolute URLs with
helper_mantis_url() and the documentation feature is deprecated, it made
more sense to have the logic in print_doc_menu() rather than creating a
new API function.

Fixes #16995

[1] https://github.com/mantisbt/mantisbt/pull/144