Shorten db_query_bound to db_query
We add a function "alias" for db_query_bound so that
existing code will continue to work.

We rename db_query_bound to db_query within mantis core.

Long term, this allows us to shorten db_query_bound() to a more sensible
db_query() call and to remove the function in 2.0 as it gives plugin
authors time to update their code and support both 1.3 and 2.0