Store 'on'/'off' filter values as boolean states
The "sticky issue" and "use date filter" filters are boolean options
that are transmitted via use of 'on' and 'off' strings in the URL.
gpc_get_bool will take these strings and return either true or false (as
a PHP bool type). We were using 'on' and 'off' within filter_api (and
elsewhere) to internally store the state of these two boolean filter
properties.

This has now been replaced with the use of a more sane true/false
representation of the state using a PHP bool type. As a consequence,
checkboxes in the filter UI throughout MantisBT now correctly reflect
the state of the filters as either on/off.