Fixes #0017783: New Event: EVENT_MANAGE_PROJECT_DELETE
Add new event: EVENT_MANAGE_PROJECT_DELETE

These events allow plugins to log or 'block' a version or project deletion,
as they occur prior to any version or project deletion.

This could be used to allow synchronisation with a 3rd party system, to deny
the deletion of specific projects/versions, or to provide auditing/event
logging.