Use New DB syntax within core
Following ad4b7bf0f31f7983e7cdb12b717c856bf2993897 which is to allow plugin
authors to make use of the new syntax within their plugins for backwards
compatiblity later on to 2.x.

We can (and should) make use of the new syntax within the MantisCore.

This will allow us to ensure that the plan to start targetting plugins for
1.3 and 2.x will be successful.