Fix marking of settings on "E-mail Notifications"
This includes multiple fixes to get the scenario working again after recent regressions.

This bug was introduced as part of  to move highlighting coloring from config to css.

- Don't generate multiple class attributes per td tag.  Introduced by 7094372255c5aa89787cc105c31c9e1f7daa649e
- Fix highlighting logic related to overriding caused by saved flags not being complete.  This is likely an old bug.
- Fix a bug where null was passed for user_id when calling config_get for global settings rather than ALL_USERS.  This could be recent, but it is very hard to tell since blame is polluted by the code style changes.
- Complete renaming 'colour' to 'color'.

I did some testing, but I wouldn't be surprised if we find more corner case bugs due to the complexity of the layers of configurations and checks involved.

Fixes #17661