mc_issue_update() email missing latest notes
The mc_issue_update() could result in updating or adding notes to the issue.  Earlier in the method the issue and issue notes are loaded and cached.  Then notes are updated.  Then email API is called which loads the issue + notes again, but this time it hits stale cache.

Fixes #17763