Admin Guide: split Configuration chapter by section
The Configuration.xml docbook file was nearly 4000 lines, making it
unwieldy to edit.

Each section was moved to a separate file in new subdirectory 'config',
and included in the main file with xi:includes.

The operation was handled using 'dbautosplit' perl script from
http://rpmfind.net/linux/rpm2html/search.php?query=dbautosplit
with some manual adjustments to the generated files.

Fixes #17781