Relationship emails disclose related issue
If a user has access to an issue and a related issue is added/deleted that the user has no access to, the user still gets a notification.

For email notifications in general, when constructing visible data checks need to take into consideration the user the email is being constructed to, rather than the current user. Since some APIs may check against current user, use override.

Fixes #9885