Allow retrieving multiple/complex configs
Add mc_config_get() to enables clients to do the following - compared to mc_config_get_string():

- Retrieve multiple configs in one call.
- Retrieve complex configs.
- Retrieve configs within context of a target user.
- Retrieve configs within context of a project.

This method returns configs in json.  I was opposing serialization, but thinking about it there is no way to do this in a strongly typed way.  Json is a good option due to the good support for decoding into native structures/objects in all languages.