Fix token upgrade error
If a user is already logged in and visits a page, then it may create a
token using the json encoding.  Then user goes to upgrade the php
unserialize() fails.  Now we check in case of php unserialize()
failure that the token isn't a valid json token before erroring out.

Fixes #17806