Don't expand reporters by default in update page
For performance reasons with update issue pages, we should revert
back to 1.2.x behavior where we don't populate the combobox of
reporters by default.  This is in order to reduce database load and
improve performance for instances with large number of reporters.

This change is also designed to make sure that search engine crawlers
don't expand such list and hence avoiding the database load.

In the future, we should use jquery auto-complete or similar control.
We could also possibly add a configuration option or cache a session
variable that determines whether to expand the list by default.  This
can be useful for instances with small number of reporters.

Fixes #17944