Fix number of displayed events lower than 50
Depending on the types of events displayed, in some cases the timeline
displayed fewer than 50 entries even though more are available.

This is because the 'Status Change' timeline events are only printed for
some status types (RESOLVED, CLOSED, REOPENED); prior to this, other
types were printed as an empty string.

This adds a new private method in IssueStatusChangeTimelineEvent class
to determine the type of status change in the constructor and store it
as a property, which is then used to reimplement the skip() method as
well as to drive html generation. This allows the timeline_events()
function to exclude unwanted entries with skip().

In addition, the logic for limiting the number of items to display is
now implemented in timeline_print_events() instead of timeline_inc.php.

Fixes #18034