Partial revert 'Additional timezone init fixes'
This reverts the changes to schema.php introduced by commit
b0d0c9dc904fa6ef6e0bc73848f216cd415429b8.

It is not necessary (and in fact, incorrect) to change timezone to UTC
prior to calling ADOConnection::BindTimestamp(), because ADOdb considers
the given timestamp to be local.

As a consequence, the timezone translation is applied twice, causing the
timestamp to be in the future.

A comment to explain this behavior has been added to the code, to avoid
this issue popping up again in the future.

Fixes #17980