Make Assign work with required custom fields
This bug was introduced in 1.3.x where users are not able to assign
issues within projects that have required custom fields on update.
This failure happens even independent of whether the
fields are populated
or not.  The fix is to enabling the Assign-To action to work independent
of custom fields values.  This this is an assign-to action rather than
a full update.

Fixes #19265