Use default category of target project when moving issues
When moving an issue having a given category to another project where
said issue does not exist, Mantis sets the issue's category according to
$g_default_category_for_moves.

However, bug_move() retrieved this setting in the context of the issue's
"old" project which is incorrect when the default category is different
in the target project (i.e. project-specific settings via Manage Config
page).

Fixes #19399