Temp fix to inability to disable notifications
The proper fix for this is not trivial involving database schema change, localization, etc.
Until this is done, we will treat uncustomizable notifications the same way as the
default preference where the "email_on_status" preference is used.
The issue will not be marked as resolved until we add the necessary preferences.
However, we should consider a model where adding user preferences doesn't
require a schema change.

Issue #19459