Protect against malicious REPORTERs
This is a proposal against enabling spammers to signup and spam
the bug tracker.  The changes limits the number of possible changes
by signed up users to 10 per hour.  Although the feature and limits are configurable.
This check doesn't apply when signup feature is not enabled or when users have access
level above the default was assigned on signup.

Issue #19261