Fix system notice in installer
When the installer is not able to connect to the database as admin, the
DB version check fails because the version cannot be retrieved from the
server. This causes display of a PHP notice. Initializing
$t_version_info fixes this problem.

Additionally, the test result is shown as 'BAD' in this case, which is
not necessarily correct (we just don't know), so we now display
'POSSIBLE PROBLEM' instead.

Fixes #19676