Consistent logic to display status legend
The code to display the status legend did not use the same logic
everywhere. We now consistently rely on binary and operator (&)
throughout the code base, like the way we handle e.g. the filter.

In particular, bug_group_action_print_bug_list() did not print the
legend when position was STATUS_LEGEND_POSITION_BOTH.

Fixes #19725