Status legend on "My View" independent of current filter
Implementation of #11554 introduced a new functionality to show just
status colors of current filter in legend, or even suppress the display
of the legend. This makes sense on "Views Issues" page but not on "My
View" page.

This introduces a new parameter for html_status_legend(), allowing the
caller to determine whether the legend should take the filter into
consideration or not (by default, it does not and will display all valid
status for the workflow).

Only "My View" page uses the new parameter.

Fixes #19573