Refactor for better api + optimize for accuracy
Refactor the API for re-usability of code between normal issue history api
and APIs used to populate the timeline functionality.  The new API also enables
only doing the minimum necessary work to render the default/full timeline view
while maintaining accuracy in terms of the number of events in the default view.