Change default page for Manage config submenu
When opened, the menu now always displays the Permissions Report page
(adm_permissions_report.php), regardless of the user's access level.

This avoids display of the Configuration Report (which was previsouly
shown by default for users having view_configuration_threshold), as that
page can be quite slow to load and moreover it is often not what admins
want to do in this menu.

Fixes #19992