Fix error when filtering on Fixed in Version
When logged in as anonmyous and trying to filter by 'Fixed in version',
Mantis throws an APPLICATION ERROR 26.

This is because the filter is initialized to default values with
filter_get_default(), which checks the filter's validity using
filter_ensure_valid_filter() where FILTER_PROPERTY_FIXED_IN_VERSION is
expected to have a type 'string' but is initialized using constant
META_FILTER_ANY which is of type 'int'.

filter_ensure_valid_filter() has been modified to set the correct type
for each property that it checks, which should fix the reported issue as
well as other occurences of incorrect data types in saved filters.

Fixes #20087