Force storing empty bugnote statistics
(1) When a bug has no bugnotes, or visible ones, force storing a 'false'
value as bugnote stats in bug cache, to not query the database again
later.

(2) Additionally, when the function bug_cache_database_result is called with
a stats array parameter, and the bug has already been cached, force the
creation of the stats index.
Previously this wasn't possible if the bug was already cached without
the bugnote stats.

Fixes #0020121