dont save filter settings for anonymous users instead of saving to DB, use session store to keep filter settings This is an attempt to fix: 0008167: Filter settings saved when using Anonymous account
This bug happens because "current" filter settings, for each project,
are saved in DB associated to the user_id of "anonymous" user,
which is unique and shared across concurrent sessions.
The proposed code hooks on the db-get and db-save functions, to detect
if user is anonymous. If so, saves (or loads) the filter string
into SESSION, not DB.
The key id for filters table, which is passed around in a cookie,
has been extended, to keep the changes minimal. If this id is negative,
means this is a key for a "current" anonymous filter, and the key value
is the project id.