CSS: Hide collapsed sections when JS is disabled
What is wrong with the collapse_api semantics:
JS writes the collapse state in a cookie; the server retrieves it to
display each section in the proper state. The classes 'collapse_closed'
and 'collapse_open' may be reversed, causing a collapsed section to have
class 'collapse_open'... wtf?

This commit fixes the problem by creating a new class to identify which
section is the collapsed one. This new class is not affected by
javascript, and hidden by default using CSS.

Fixes #19508

Signed-off-by: Damien Regad <dregad@mantisbt.org>

Changes from cproensa's original commits:
- squashed
- commit message reworded
- CSS rule on a single line