Remove href on collapse-link when JS is disabled
When JS is disabled, the collapsible section is always shown open; the
'+' icon is therefore useless, and it does not make sense to turn it
into a hyperlink.

The href attribute has been removed (the existing javascript handles
adding the hyperlink when enabled), and the cursor is set to 'pointer'
via javascript.