collapse_api: hide icon and remove extra spaces
When javascript is disabled, it does not make sense to display the
collapse icon. This commit hides it by default, and lets the JS show it.

Also remove the &nbsp; after the icon, replace it with CSS padding, and
adjust the generated HTML so that it does not generate extra spaces.

Note: a &nbsp; was added to filter_draw_selection_area2(), as a hack to
ensure the filter div is high enough to fit the search field and
buttons.