Create db helper function to generate IN clauses
Add a helper function to format an sql IN clause,
passing an array as parameter.
This will return a string, with prepared db_param
placeholders, as many as elements in the array.
If the array contains only one element, return
a simple "equals" operation.

This can improve many parts of code where IN clauses
are generated as an array implode without using
db_param.