Improve print_user_option_list memory usage
Use directly the query result to save on intermediate
memory storage for the full user array.
This is possible because the query is already sorted
by system defaults.
However, when sort by last name is applicable, we have to
fetch the whole data first, because last name is not
a database field, thus the query can't be sorted on it.