Fix resetting of access level when updating workflow
This is the same fix as 63bbb19393c3a2b9bdc7e159992c160561a7ce09, and
addresses the incorrect behavior for the fields
- Who can change workflow
- Who can change access levels
by applying the same mechanism used to detect a change in the status to
the access level, i.e.  comparing the updated access level with the
parent and not the current value.

Fixes #16771

Signed-off-by: Damien Regad <dregad@mantisbt.org>

Original commit message reworded