Check workflow prior to reopening bug
Add a check to access_can_reopen_bug() to ensure the reopen status is
reachable by the current workflow configuration. If not, return false.

Previously, the reopen operation would fail anyway when the status is
changed. With this early check, failure is detected before the action is
made.

A direct benefit from this is that the "reopen" button is not shown to
the user if the action is not possible.

Fixes #20256

Signed-off-by: Damien Regad <dregad@mantisbt.org>