Remove call to project_includes_user
project_includes_user() is called once for every user that is displayed
in project manage page, to check if user is directly assigned to the
project and print the "remove" button.  This function performs a db
query for each user, which degrades performance as user count grows, and
is especilally severe on public projects.

Instead, fetch all local user at once and compare against the saved
array if user is directly assigned to the project.

Fixes #20242