Suppress unserialize() errors when converting tokens to json
install_check_token_serialization() converts existing tokens from php
serialization to json.

It may happen that the DB already contains tokens in json format, in
which case depending on error settings, a PHP notice can be shown to the
user which is confusing because the upgrade step's status is GOOD and
the upgrade completes successfully.

Fixes #20326