Fix 2-byte attachment saved to DB when upload_method is DISK
Commit 4f4e69bd73e2840ce4542072bb544f51ddaf3cf7 changed file_add() to
use bind parameters, but failed to remove quotes escaping of $c_content
variable.

This results in an invalid 2-byte attachment (containing 2 single
quotes "''") being stored in the file table in addition to the one saved
to disk as expected, which does not actually have any impact unless the
admin tries to move attachments from DB to disk; in that case, the valid
files are overwritten, causing loss of data.

The problem is fixed by setting $c_content to an empty string.

Previously fixed in master branch commit
02bbb99501c48f89ad249fb505dc9bba93bae5c0

Fixes #20340