CSS class renamed from 'has-required' to 'required'
There was an inconsistency in the class name used to flag mandatory
fields. The fields use 'required' class to display a red asterisk next
to the label, while the parent fieldset relied on 'has-required' to
display the '* required' legend at the bottom of the form.

This caused password fields in account_page.php (which are only required
in specific contexts, e.g.  when changing a new account's password after
verification) not to be flagged with the red asterisk.

We now consistently use 'required'.

Fixes #20395