Bugnote count optimizations
Some optimizatios are done to bugnote stats
functions, to avoid redundant queries.

- Allow storing 'null' as _stats value.
Previously, if the stats register is not existant,
it because either it hasn't been calculated yet,
or the bug has no notes.
Now, if the stats register is explicitly null, the
stats has been calculated and we know bug has no notes.

- Remove check for bug access when building stats
It was not directly needed, and has a conflict with
the bug cache build process:
The stats are calcualted before creating the bug cache,
but getting access info for individual bug relies in
bug cache. As bug cache is empty, it would trigger
queries for each bug info.
This does not affect note visibility checks, because
its a configuation at project level, not bug level.

- Fix warnings
- Fix comments