Change MySQL default charset and collation for new DB
If MySQL version >= 5.5.3, we use the 'utf8mb4' character set instead of
'utf8'. This allows storage of up to 4-byte unicode characters (utf8
only allows 3 bytes, so not all characters can be represented [1].

Collation changes from 'general_ci' to 'unicode_ci', because as
explained in MySQL manual [2]: "comparisons for the utf8_general_ci
collation are faster, but slightly less correct, than comparisons for
utf8_unicode_ci. The reason for this is that utf8_unicode_ci supports
mappings such as expansions; that is, when one character compares as
equal to combinations of other characters."

In the future, we should probably let the admin update these settings as
part of the installation process.

Fixes #20431

[1] https://dev.mysql.com/doc/refman/5.7/en/charset-unicode-utf8mb4.html
[2] http://dev.mysql.com/doc/refman/5.7/en/charset-unicode-sets.html