Fix gpc_get_xxx_array for associative arrays
Functions:
- gpc_get_int_array(),
- gpc_get_bool_array()
dont work properly for html fields named with associative indexes
(numeric out or order, or strings), because it assumes indexes as
zero-based sequential numbers.

The code is changed to keep the original keys as defined in the field
names.

This is not an issue in MantisBT core, but may affects plugins.

Fixes #20476

Signed-off-by: Damien Regad <dregad@mantisbt.org>