Fix pgsql query error manage_user_page
Due to pgsql having positional parameters in parametrized queries, the
page is throwing  an error in some combination of conditions.

The error is caused by the reutilization of a prepared WHERE condition
in a string from previous query. In the new query, a db_param() call is
issued, which has the parameter reset to 1 (returning "$1").  Then,
the resulting query string contains two "$1" parameters for different
values.

With this commit, the query build code is cleaned to avoid said
situation.

Fixes #20483

Signed-off-by: Damien Regad <dregad@mantisbt.org>