Consistently push query params in all APIs
Prior to building a database query, our APIs must ensure that they leave
the parameters numbering in the same state as it was prior to the
function call.

While failure to do so does not have any impact with MySQL, it causes
errors with RDBMS which rely on strict parameter ordering (PostgreSQL
and Oracle) when a query is executed while another is being built,
because the parameter count is reset.

All MantisBT APIs have been reviewed, and db_push_param() calls added
where needed to ensure that the caller always finds the environment in
the same state as it was prior the API call.

A new db_pop_param() API was added to cover rare cases where a query is
called multiple times, or built but never executed.

Fixes #20479