Further fix parameter count for pgsql
We need an internal parameter push/pop to keep
track of parameter count when using pgsql.
This is due to ADOdb returning token numbered by
its internal counter, which is resetted al query execution.
Our inner param pop restores that count in ADOdb.

The internal push/pop is needed anyway for the '%' substitutions
which also should not interfere with an upper level sql build
using parameters, even in the situation that db_param was not
explitly used for current sql, but is needed to replace
the '%' tokens.