Fix passing parameters for EVENT_TYPE_OUTPUT events
Some events of EVENT_TYPE_OUTPUT specify a parameter
to be passed to the callback function.
However the event dispatcher for that type was not
designed to receive a parameter in that way.

Modified the event function to allow for that parameter
to be passed on, thus matching documented specification
and current use in code base.

Affected events are:
EVENT_MANAGE_OVERVIEW_INFO
EVENT_VIEW_BUG_ATTACHMENT

Those events have not been working properly until now.

Fixes #20535