Remove the "new syntax" for query parameters
This was introduced in ad4b7bf0f31f7983e7cdb12b717c856bf2993897 but is
currently not used anywhere in the code.

Considering the potential for problems this introduces right now with
interleaved queries and PostgreSQL, it's better to revert now as
discussed [1], and put it back later if we decide to do so.

[1] https://github.com/mantisbt/mantisbt/pull/705#issuecomment-172260207