Dont use serialization for Session variables
There is no need to serialize data inserted in $_SESSION, as php does it
at the end of script

This is a performance hit when a session key is used multiple times (eg:
'form_security_tokens')

Additionally, we use a new identifier for the session key (by adding a
version number to it). Since we changed the format of the stored data,
this will ignore older session data to avoid errors for users having
active sessions during the transition.

Fixes #20142, #18016

Signed-off-by: Damien Regad <dregad@mantisbt.org>

Squashed original commits and reworded commit message