Set random, not empty, pass on reset if no email
If password reset is not going to send out reset email
(e.g. because email sending is disabled)
then instead of setting empty password "",
we will set a random password,
and show it to the admin who is performing the reset.

Secure and comfortable.

This code was tried by me and it works in the usecase above.
It was NOT tested in other case (where e-mail should be sent),
and the code was NOT reviewed regarding security.

Just sending the patch so other developers can finish it as they see fit.