Random password reset: information message update
To show before resetting password to random

English and Polish is actually translated,
for other languages it constructs a simple message
from available translated words.

This code was tried by me and it works in the usecase above.
It was NOT tested in other case (where e-mail should be sent),
and the code was NOT reviewed regarding security.

Just sending the patch so other developers can finish it as they see fit.